#3010 minor modal changes
